Website/webserver/src/requirements.txt
pyramid==1.10
pyramid-jinja2
mysql-connector-python
python-dotenv
dronekit
dronekit_sitl
serial
pyserial
requests==2.18